requirements.txt
os
flask
shutil
google-genai